mybackend/requirements.txt
Django==5.0.6
oauthlib==3.2.0
requests==2.25.1
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.3.6
django-cors-headers==4.3.1
django-countries==7.6.1
djangorestframework_simplejwt==5.3.1
async-timeout==4.0.3
channels==3.0.5
channels-redis==4.2.0
msgpack==1.0.8
redis==5.0.6
whitenoise==6.6.0



